feat(armies): Split script to fill tables
