+++ scripts/armies_database/src/global_strings.py
@@ -1,0 +1,84 @@
+REPOSITORY_NAME: str = "The-9th-Age"
+
+NAME: str = "name"
+VALUE: str = "value"
+TARGET_ID: str = "targetid"
+TYPE: str = "type"
+
+MODIFIERS: str = "modifiers"
+MODIFIER: str = "modifier"
+
+FIELD: str = "field"
+
+PERCENT_VALUE: str = "percentvalue"
+
+DESCRIPTION: str = "description"
+
+CONDITIONS: str = "conditions"
+CONDITION: str = "condition"
+
+ID: str = "id"
+
+CONSTRAINTS: str = "constraints"
+CONSTRAINT: str = "constraint"
+
+RULES: str = "rules"
+RULE: str = "rule"
+
+COSTS: str = "costs"
+COST: str = "cost"
+
+INFOLINKS: str = "infolinks"
+INFOLINK: str = "infolink"
+
+SELECTION_ENTRIES: str = "selectionentries"
+SELECTION_ENTRY: str = "selectionentry"
+
+ENTRY_LINKS: str = "entrylinks"
+ENTRY_LINK: str = "entrylink"
+
+COLLECTIVE: str = "collective"
+
+SELECTION_ENTRY_GROUPS: str = "selectionentrygroups"
+SELECTION_ENTRY_GROUP: str = "selectionentrygroup"
+
+TYPENAME: str = "typename"
+
+CHARACTERISTICS: str = "characteristics"
+CHARACTERISTIC: str = "characteristic"
+
+PROFILES: str = "profiles"
+PROFILE: str = "profile"
+
+CATEGORY_LINKS: str = "categorylinks"
+CATEGORY_LINK: str = "categorylink"
+
+CATALOGUE: str = "catalogue"
+
+SHARED_SELECTION_ENTRIES: str = "sharedselectionentries"
+
+FORCE_ENTRIES: str = "forceentries"
+FORCE_ENTRY: str = "forceentry"
+
+SHARED_SELECTION_ENTRY_GROUPS: str = "sharedselectionentrygroups"
+
+SHARED_RULES: str = "sharedrules"
+
+SHARED_PROFILES: str = "sharedprofiles"
+
+
+ARMIES_TABLE: str = "armies"
+RULES_TABLE: str = "rules"
+UNIT_CATEGORIES_TABLE: str = "unit_categories"
+UNIT_PROFILES_TABLE: str = "unit_profiles"
+UPGRADE_CATEGORIES_TABLE: str = "upgrade_categories"
+UNITS_TABLE: str = "units"
+OPTIONS_TABLE: str = "options"
+MODIFIERS_TABLE: str = "modifiers"
+UPGRADES_TABLE: str = "upgrades"
+SPECIAL_ITEMS_TABLE: str = "special_items"
+SPECIAL_ITEM_CATEGORIES_TABLE: str = "special_item_categories"
+
+ARMY: str = "army"
+IS_SHARED: str = "is_shared"
+OWNER: str = "owner"
